move test & add check for results
