+++ pysb/tests/test_integrate.py
@@ -3,7 +3,7 @@ from ast import Param
 from pysb.testing import *
 from pysb.integrate import Solver
 import numpy as np
-from pysb import Monomer, Parameter, Initial, Observable, Rule, Expression, time
+from pysb import Monomer, Parameter, Initial, Observable, Rule, Expression
 from pysb.bng import run_ssa
 from pysb.simulator.base import SimulatorException
 
@@ -119,21 +119,6 @@ def test_integrate_with_expression():
     assert solver.yexpr_view.shape == (len(tspan),
                                        len(model.expressions_dynamic()))
     assert solver.yobs_view.shape == (len(tspan), len(model.observables))
-
-@with_model
-def test_integrate_with_time():
-    Monomer('A')
-
-    Parameter('ksynthA', 1)
-
-    Expression('synthrate', ksynthA*time)
-
-    Rule('synthesize_A', None >> A(), synthrate)
-
-    tspan = np.linspace(0, 40)
-
-    solver = Solver(model, tspan)
-    solver.run()
 
 
 def test_robertson_integration():
